Connect and discovering hosts done
The screens all work, and if other devices were broadcasting we could
find them.  Creates the dynamic radio buttons based on what devices are
around, you select one and click the connect button and it runs the
connect function from the client system to connect to the device and
proceed to main screen. Search runs the discovery method and updates
the list.